Move permission related methods into a separate extension
