+++ Source/Model/User/ZMUser+Teams.swift
@@ -27,85 +27,9 @@ public extension ZMUser {
     @objc var team: Team? {
         return membership?.team
     }
-    
-    var teamRole: TeamRole {
-        return TeamRole(rawPermissions: permissions?.rawValue ?? 0)
-    }
-    
-    private var permissions: Permissions? {
-        return membership?.permissions
-    }
-    
+        
     @objc static func keyPathsForValuesAffectingTeam() -> Set<String> {
          return [#keyPath(ZMUser.membership)]
-    }
-
-    @objc(canAddUserToConversation:)
-    func canAddUser(to conversation: ZMConversation) -> Bool {
-        guard conversation.teamRemoteIdentifier == nil || !isGuest(in: conversation), conversation.isSelfAnActiveMember else { return false }
-        return permissions?.contains(.addRemoveConversationMember) ?? true
-    }
-
-    @objc(canRemoveUserFromConversation:)
-    func canRemoveUser(from conversation: ZMConversation) -> Bool {
-        guard !isGuest(in: conversation), conversation.isSelfAnActiveMember else { return false }
-        return permissions?.contains(.addRemoveConversationMember) ?? true
-    }
-    
-    @objc(canModifyReadReceiptSettingsInConversation:)
-    func canModifyReadReceiptSettings(in conversation: ZMConversation) -> Bool {
-        guard !isGuest(in: conversation), conversation.isSelfAnActiveMember else { return false }
-        return permissions?.contains(.modifyConversationMetaData) ?? false
-    }
-    
-    @objc(canModifyEphemeralSettingsInConversation:)
-    func canModifyEphemeralSettings(in conversation: ZMConversation) -> Bool {
-        guard  conversation.teamRemoteIdentifier == nil || !isGuest(in: conversation), conversation.isSelfAnActiveMember else { return false }
-        return permissions?.contains(.modifyConversationMetaData) ?? true
-    }
-    
-    @objc(canModifyNotificationSettingsInConversation:)
-    func canModifyNotificationSettings(in conversation: ZMConversation) -> Bool {
-        guard conversation.isSelfAnActiveMember else { return false }
-        return isTeamMember
-    }
-    
-    @objc(canModifyAccessControlSettingsInConversation:)
-    func canModifyAccessControlSettings(in conversation: ZMConversation) -> Bool {
-        guard !isGuest(in: conversation), conversation.isSelfAnActiveMember else { return false }
-        return permissions?.contains(.modifyConversationMetaData) ?? false
-    }
-    
-    @objc(canModifyTitleInConversation:)
-    func canModifyTitle(in conversation: ZMConversation) -> Bool {
-        guard conversation.isSelfAnActiveMember else { return false }
-        return permissions?.contains(.modifyConversationMetaData) ?? true
-    }
-
-    @objc var canCreateConversation: Bool {
-        return permissions?.contains(.createConversation) ?? true
-    }
-
-    @objc func _isGuest(in conversation: ZMConversation) -> Bool {
-        if isSelfUser {
-            // In case the self user is a guest in a team conversation, the backend will
-            // return a 404 when fetching said team and we will delete the team.
-            // We store the teamRemoteIdentifier of the team to check if we don't have a local team,
-            // but received a teamId in the conversation payload, which means we are a guest in the conversation.
-            
-            if let team = team {
-                // If the self user belongs to a team he/she's a guest in every non team conversation
-                return conversation.teamRemoteIdentifier != team.remoteIdentifier
-            } else {
-                // If the self user doesn't belong to a team he/she's a guest in all team conversations
-                return conversation.teamRemoteIdentifier != nil
-            }
-        } else {
-            return !isServiceUser // Bots are never guests
-                && ZMUser.selfUser(in: managedObjectContext!).hasTeam // There can't be guests in a team that doesn't exist
-                && conversation.lastServerSyncedActiveParticipants.contains(self)
-                && membership == nil
-        }
     }
 
     @objc var isWirelessUser: Bool {
