+++ Source/Model/User/ZMUser.swift
@@ -27,18 +27,6 @@ extension ZMUser: UserType {
     public func isGuest(in conversation: ZMConversation) -> Bool {
         return _isGuest(in: conversation)
     }
-
-    public func canAccessCompanyInformation(of user: UserType) -> Bool {
-        guard
-            let otherUser = user as? ZMUser,
-            let otherUserTeamID = otherUser.team?.remoteIdentifier,
-            let selfUserTeamID = self.team?.remoteIdentifier
-        else {
-            return false
-        }
-        
-        return selfUserTeamID == otherUserTeamID
-    }
     
     public var teamName: String? {
         return team?.name
